LWDEV-6992 add convertor for fee instructions

+++ src/main/kotlin/com/lykke/matching/engine/daos/FeeInstruction.kt
@@ -3,12 +3,11 @@ package com.lykke.matching.engine.daos
 import com.lykke.matching.engine.daos.fee.NewFeeInstruction
 import com.lykke.matching.engine.messages.ProtocolMessages
 import java.io.Serializable
-import java.math.BigDecimal
 
 open class FeeInstruction(
         val type: FeeType,
         val sizeType: FeeSizeType?,
-        val size: BigDecimal?,
+        val size: Double?,
         val sourceClientId: String?,
         val targetClientId: String?
 ) : Serializable {
@@ -26,7 +25,7 @@ open class FeeInstruction(
             return FeeInstruction(
                     feeType,
                     sizeType,
-                    if (fee.hasSize()) BigDecimal.valueOf(fee.size)  else null,
+                    if (fee.hasSize()) fee.size else null,
                     if (fee.hasSourceClientId()) fee.sourceClientId else null,
                     if (fee.hasTargetClientId()) fee.targetClientId else null
             )
