+++ src/main/kotlin/com/lykke/matching/engine/daos/fee/NewFeeInstruction.kt
@@ -4,11 +4,10 @@ import com.lykke.matching.engine.daos.FeeSizeType
 import com.lykke.matching.engine.daos.FeeType
 import com.lykke.matching.engine.messages.ProtocolMessages
 import com.lykke.matching.engine.daos.FeeInstruction
-import java.math.BigDecimal
 
 open class NewFeeInstruction(type: FeeType,
                              takerSizeType: FeeSizeType?,
-                             takerSize: BigDecimal?,
+                             takerSize: Double?,
                              sourceClientId: String?,
                              targetClientId: String?,
                              val assetIds: List<String>) : FeeInstruction(type, takerSizeType, takerSize, sourceClientId, targetClientId) {
@@ -27,7 +26,7 @@ open class NewFeeInstruction(type: FeeType,
             return NewFeeInstruction(
                     feeType,
                     sizeType,
-                    if (fee.hasSize()) BigDecimal.valueOf(fee.size) else null,
+                    if (fee.hasSize()) fee.size else null,
                     if (fee.hasSourceClientId()) fee.sourceClientId else null,
                     if (fee.hasTargetClientId()) fee.targetClientId else null,
                     fee.assetIdList.toList()
